+++ src/Commands.cs
@@ -29,7 +29,7 @@ namespace CwaffingTheGungy
             ETGModConsole.Commands.AddGroup("gg", delegate (string[] args)
             {
                 LootEngine.SpawnItem(
-                    PickupObjectDatabase.GetById(IDs.Pickups["deadline"]).gameObject,
+                    PickupObjectDatabase.GetById(IDs.Pickups["aimu_hakurei"]).gameObject,
                     GameManager.Instance.PrimaryPlayer.CenterPosition,
                     Vector2.zero,
                     0);